refactor: adjust according to review
