+++ src/components/app/userActionFormSuccessScreen/UserActionFormSuccessScreenFeedback.tsx
@@ -19,7 +19,7 @@ export function UserActionFormSuccessScreenFeedback(
   const { image, title, isVotingDay } = props
   const description = isVotingDay
     ? props.description // Temporary description until 2024 election is over.
-    : 'This year’s election is critical to the future of crypto. Here’s how to become an informed voter.'
+    : "This year's election is critical to the future of crypto. Here's how to become an informed voter."
 
   return (
     <div className="flex flex-col items-center justify-center gap-8">
