+++ src/components/app/userActionVotingDay/sections/success.tsx
@@ -1,26 +1,15 @@
 import { USER_ACTION_FORM_SUCCESS_SCREEN_INFO } from '@/components/app/userActionFormSuccessScreen/constants'
 import { UserActionFormSuccessScreenFeedback } from '@/components/app/userActionFormSuccessScreen/UserActionFormSuccessScreenFeedback'
+import { DidYouVoteVideo } from '@/components/app/userActionVotingDay/didYouVoteVideo'
 import { Button } from '@/components/ui/button'
-import { NextImage } from '@/components/ui/image'
 import { ExternalLink } from '@/components/ui/link'
-import { NFTSlug } from '@/utils/shared/nft'
 import { createTweetLink } from '@/utils/web/createTweetLink'
-import { NFT_CLIENT_METADATA } from '@/utils/web/nft'
-
-const I_VOTED_NFT_IMAGE = NFT_CLIENT_METADATA[NFTSlug.I_VOTED].image
 
 export const UserActionIVotedSuccess = () => {
   return (
     <div className="flex h-full flex-col items-center justify-between gap-8">
       <UserActionFormSuccessScreenFeedback
-        image={
-          <NextImage
-            alt={I_VOTED_NFT_IMAGE.alt}
-            height={120}
-            src="/actionTypeIcons/iVoted.png"
-            width={120}
-          />
-        }
+        image={<DidYouVoteVideo />}
         isVotingDay
         {...USER_ACTION_FORM_SUCCESS_SCREEN_INFO.VOTING_DAY}
       />
